+++ src/libcsg/modules/io/h5mdtrajectoryreader.cc
@@ -193,7 +193,7 @@ void H5MDTrajectoryReader::Initialize(CSG_Topology &top) {
   // TODO: reads mass, charge and particle type.
 
   if (has_id_group_ == H5MDTrajectoryReader::NONE && top.BeadCount() > 0 &&
-      N_particles_ != top.BeadCount()) {
+      static_cast<size_t>(N_particles_) != top.BeadCount()) {
     cout << "Warning: The number of beads (" << N_particles_ << ")";
     cout << " in the trajectory is different than defined in the topology ("
          << top.BeadCount() << ")" << endl;
